Refactor bottlerocket into its own daemon
* can't currently cross-compile cgo apps using bazel
  - this allows domoticsd to be cross compiled until solved
