+++ services/domotics/bridge/BUILD.bazel
@@ -16,7 +16,6 @@ go_library(
         "//vendor/github.com/golang/protobuf/proto:go_default_library",
         "//vendor/github.com/google/uuid:go_default_library",
         "//vendor/github.com/mattn/go-sqlite3:go_default_library",
-        "//vendor/github.com/rmrobinson/bottlerocket-go:go_default_library",
         "//vendor/github.com/rmrobinson/hue-go:go_default_library",
         "//vendor/github.com/rmrobinson/monoprice-amp-go:go_default_library",
     ],
