+++ services/domotics/cmd/bottlerocketd/BUILD.bazel
@@ -3,27 +3,25 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "hue.go",
+        "bottlerocket.go",
         "main.go",
-        "monopamp.go",
-        "proxy.go",
     ],
-    importpath = "github.com/rmrobinson/nerves/services/domotics/cmd/domoticsd",
+    importpath = "github.com/rmrobinson/nerves/services/domotics/cmd/bottlerocketd",
     visibility = ["//visibility:private"],
     deps = [
+        "//bazel-nerves/services/domotics/bridge:go_default_library",
         "//services/domotics:go_default_library",
         "//services/domotics/bridge:go_default_library",
-        "//vendor/github.com/rmrobinson/hue-go:go_default_library",
-        "//vendor/github.com/rmrobinson/monoprice-amp-go:go_default_library",
+        "//vendor/github.com/golang/protobuf/proto:go_default_library",
+        "//vendor/github.com/rmrobinson/bottlerocket-go:go_default_library",
         "//vendor/github.com/spf13/viper:go_default_library",
-        "//vendor/github.com/tarm/serial:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
     ],
 )
 
 go_binary(
-    name = "domoticsd",
+    name = "bottlerocketd",
     embed = [":go_default_library"],
     visibility = ["//visibility:public"],
 )
