+++ services/news/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
     importpath = "github.com/rmrobinson/nerves/services/news",
     visibility = ["//visibility:public"],
     deps = [
+        "//vendor/github.com/golang/protobuf/proto:go_default_library",
         "//vendor/github.com/mmcdole/gofeed:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
         "//vendor/golang.org/x/net/html:go_default_library",
